requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.3
psycopg2-binary==2.9.10
pydantic==2.9.2
ollama==0.1.7
pgvector==0.2.5
